Print, math expression and validation
