+++ utils.py
@@ -1,0 +1,8 @@
+
+
+def validate_input(var_name):
+    while True:
+        try:
+            return int(input('{}:'.format(var_name)))
+        except ValueError:
+            print('Please, enter a NUMBER, retarded')
